Add DMMT5551 symbol definition
So the tremolo board is entirely useless as a guitar pedal, but I've
mentioned several times that it is probably my favorite board for its
intended purpose: learning about analog circuits.

However, I picked entirely the wrong dual transistor in the MBT2222ADW1,
and this is my replacement component. Let's see if it's any better.

The MTB2222 was a fine general-purpose transistor, and it's not like it
was a horrible choice, but it ended up being a bit of a pain.

The main goal here is to avoid the soldering pain of the small SOT-363 /
SC-88 / SC70-6 package (don't you love SMD package naming? "The good
thing about standards is that there are so many to choose from").

It actually hasn't been a huge problem - the SOT-363 is still much
better than the horrid DFN2x2 that is the HUSB237. So unlike the DFN
package, I think I have a 100% success rate at soldering those things
(knock wood), but I still don't enjoy it.

So let's learn from our mistakes, and pick a SOT-23-6 (SOT26 / SC59-6 /
TSOP-6 / whatever) package instead. These things are pennies, I'd rather
expand on my component library than pointlessly stick to my first
choice. And it's not like I am space constrained.

The secondary goal here is that I think I want to learn about current
sources and mirrors, so while initially I'll just replace the footprint
for the dual transistor that I used for the oscillator on the tremolo
board, let's go for something that is documented to be a matched pair
("2% Matched Tolerance").

The DMMT5551 also has a PNP version in the DMT5401, I'll do that
separately.

Now, I don't think that matching really matters for anything I will ever
do, but it still feels like the right thing to do. And I don't think any
of the other differences between these transistors really matters either
(like the CE breakdown voltage of 40V vs 160V).

Do I know what I'm doing? No. This might not be a great choice either.
But that's kind of the point. "Learning experience".

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>