Fix up DMMT5401 symbol
When I was starting to use this symbol, I ended up messing up several
times just because I had drawn this symbol oddly.

So move the emitter to the top, so that the arrow points down, and the
default layout is the sane one and you don't need to mirror the thing
every time (or forget to mirror it, which then causes confusion).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>