Add current mirror test / oscillator circuit
This is an attempt at a better oscillator circuit that gives both a
clean triangle wave and square wave output.

The triange wave comes from a op-amp integrator that is driven by a
current source.

The square wave comes from a schmit trigger comparator driven by the
triangle wave, which then in turn also drives the transistor that
switches the direction of the current for the comparator.

This does seem to work in simulation.  And I did enough of the PCB
layout to see roughly how much space it takes.  It's obviously rather
much bigger than the simple two-transistor oscillator I used for the
tremolo, but it's not too bad.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>