Connect up oscillator into a test board
Make a 'sineish' output by using a dual opamp and another integrator
circuit to turn the triangle wave into something that approximates a
sine wave.

It would be closer if my triangle wave was more balanced, but it looks
reasonable.  Even if the volume ends up being frequency dependent
because of how the integrator works - I haven't figured out how to fix
that part up.

Now it's an "effects pedal" with no inputs, only an output.  Simplest
way to test it.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>