What's better than a rectifier? *Three* rectifiers
Ok, this is just for fun, but since I had the test boards built from
splitting up the original Octaver circuit, I put three of them in series
just because I wanted to see what that would do.

And it sounded kind of interesting.  It also looks cool on the 'scope.

So I'm building the darn thing, because I can.  I should probably add a
switch to pick between doing two or three rounds of rectification, and
maybe a volume pot or whatever, but I started just laying the three
rectification circuits out in series and simulating it, and this is the
result of that.

And honestly, three rounds of rectification sounded more interesting
than two.  Or one.  Why have a switch for "boring"?

It looks as fun in simulation as it looked on the scope.

Of course, this is another of those "it's entirely useless from a
musical standpoint", but I've gotten over worrying about whether
anything I do is useful or not. I've got real work for _that_.

The tremolo and signal generator pedal sure aren't useful, but they were
educational.

This isn't even educational, this is just silly.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>