Add blend circuit to the triple rectifier
I "tested" this circuit in real life by just chaining three rectifiers
together, along with a splitter/joiner board.  It sounds ok.  But it
definitely wants the blending circuit, and having that splitter/joiner
just for that is very inconvenient.

So just make it one single pedal so that it's actually useful (not that
it's _useful_ useful, if you get what I mean..).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>