Add simple model for quick test shells
For the pure rectifier pedal I wanted to have something simpler for
testing than a clone of the actual real Hammond enclosures.

In particular, I didn't want to deal with screws and tight fitting
shells, I just want something really simple that keeps the boards held
together and avoids any issues with touching the electronics that then
causes noise issues.

The solution is zip-ties.  With a very simple "lid" on both sides that
is quick and easy to print and doesn't even attempt to look like
anything but what it is: a temporary test setup.  Good for throw-away
pedal test circuits.

This one is designed for not even having any controls (like that
rectifier pedal), but due to the simplicity it's also really easy to
just mess with the suit whatever needs.

"Functional, and simple"

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>