Add a few symbol and footprint definitions for new power supply trial
This is a LT8330 using a coupled inductor ("Ćuk Converter") along with
the Schottky diode suggested by the datasheet.

The Ćuk topology _should_ be less noisy, and the switching frequency of
the LT8330 is 2MHz, so hopefully any noise is not audible.  But if this
mostly works, I foresee several rounds of "let's try to fix the
switching noise".

I'm trying the dual rails again because I have a cunning plan for a new
way to connect the effect board and base board, and if I re-do that, I
really would like to have better power.

To offset the whole power supply complexity, this will obviate the need
for the virtual ground, and I also just use a single diode for the
reverse polarity protection (since the voltage rail is much wider and
the diode drop isn't very noticeable as a result).

And the LT8330 has an integrated switch and doesn't need all that much
in the form of external components.  So it's not actually _that_ much
more complex than the TLE2426 VGND model was.

Apart from the potential noise issues, of course.  Let's see.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>