Add trial modular stomp switch setup with dual voltage rails
This is a quick "let's test this" thing where the stomp switch and power
supply are actually physically separated from the effects board, using a
simple ethernet cable to connect both power and signals.

If this ends up being workable, I'll redo the enclosures and layouts
etc, because that changes the whole form factor model.

This feels like the solution to "what is the proper connection between
the base board and effect board".  And since the connector now changes
fundamentally, let's try - once again - doing dual power rails.  This
time using a Ćuk converter (with two inductors, although done as one
coupled inductor component).

An even better design would use actual differential signals, but this is
just doing single-ended "signal and ground" twisted pairs for both
directions (so two pairs of two wires each), and then using the other
two pairs of wires for the two voltage rails.

That's very wasteful of wires, but it's what "mode B" PoE cables (and
various hacky non-standard vendor-specific PoE setups) did too.  So
there's a solid / sordid history of that kind of wasteful wire use.

But *if* we want any other kind of signalling, we're all out of wires.
So even for just LEDs on the effect board, we don't have any way to even
say "this board is active".

I _could_ try to use real differential signalling on top of the DC
offset that is the power rails (like real "modern" PoE does).  But that
feels like it would complicate the whole effect box too much to be worth
it.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>