Fix up component size on oscillator board
When actually receiving the boards and starting to build them up, I
noticed that I made a couple of the small caps be 1206 sized.  I do have
those, but my standard size for the nF caps is 0805, and that's what all
my small C0G caps are.

So fix it up for if I ever re-use this.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>