Add notes about oscillator test results
Short answer: yes, it matches simulation and seems to work exactly as
expected.

Useful? As a way to learn about current sources it was great.  As a
guitar pedal, it does away with the need for a guitar, and makes for a
very loud triangle wave.  But since I don't actually do synths, I think
the real use-case might be with the triangle wave generator as part of
an improved tremolo circuit or whatever.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>