Do a much better job at re-biasing the input for the Gilbert cell
So in commit f2789ed ("Commit random multiplier state") I made the
multiplier work much better by properly re-biasing the input for the
lower half of the Gilbert cell, but I didn't do a very good job at it.

I left this comment about how I really needed to make a lower-impedance
version of the hacky re-biasing I did.

This is that lower impedance version.  I may have gone a bit overboard
in using both op-amps of the OPA1656 to do this: one makes the voltage
reference rail very solid (it's ~2.8V - not a significant number other
than it's just in the right ballpark and easy to generate with standard
resistors).

The other op-amp then just buffers the scaled and rebased input signal.

The end result of this is that the bottom differential amplifier output
is a _lot_ more solid, which in turn means that the overall multiplier
output looks a lot better and more balanced.  And with the scaling of
the that second input signal, we should have reasonable control over the
overall resulting volume level.

Possibly the original input signal should also be scaled just to get
better linearity? Depending on the use case (like signal squaring),
distortion can be kind of the point, so that will depend on what the end
result actually does.

And a real effects pedal would typically need a blending stage too.  I
guess I will get to that at some point.

It would be good to get some kind of verification from some HW person
that this is actually looking mostly sane, but it does seem to at least
simulate well, and it matches my current (admittedly recent and probably
fairly limited) understanding of the Gilbert cell.

Note: this uses a ridiculous simulation function for the control voltage
just to verify that "yes, the end result really looks like a properly
scaled four-quadrant multiply".

In a real pedal, the control voltage would either be a second scaled
version of the input (for an octaver that uses a squaring function
rather than "abs()"), or be some kind of signal level function (for a
compressor), or be a separate signal generator thing (eg for a tremolo
or other).

So this particular thing is right now just a generic Gilbert cell
multiplication thing.  I will have to think about how I actually use it
to test it out properly.  Because I do want to test that my
understanding actually translates to reality too, not just simulation.

Maybe I just will need to make a two-input pedal?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>