Add ImSquared effect where we use the multiplier as an octaver
It basically squares the input by multiplying it with itself.  With some
scaling and blending pots to play with it.

This feels like it might be better than the rectifier, in how it avoids
the harsh edges when crossing the zero point.  But so far my judgement
on how it sounds from looking at the signal hasn't been very good, so...

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>