Use better signals for simulation
I really need to learn not to use a sine signal for simulation input
testing: it makes the output look artificially good and hides problems.

I had fixed that problem for the control voltage in the multiplier
circuit, but then never did it for the main input signal, because the
end result looked so "obviously correct".

But then I got suspicious, and sure enough, when I start using a more
interesting main signal the picture changes.  It's not doing a proper
multiplication in the ImSquared circuit.  Maybe that might be ok for
sound, but it's not what I wanted the circuit to actually do.

Yes, VSIN is really easy to use, and it's a pain to set up something
even as simple as a triangle wave, and ngspice seems to get unhappier
about abrupt edges so simulation is more prone to problems.

But making a signal that is more triangle-wave like is not _that_ hard,
and at least shows when the end result has soft clipping or other simple
issues, and sure enough...

Note that the multiplier itself seems ok.  The problem is in the
ImSquared circuit, and I think I see what it is: blending in the
original signal - even when it's not used - messes up the impedance of
the input signal to the multiplier, and I may just need to add a buffer
there too.

This doesn't do that fix, this just fixes the simulation input so that
you can see the problem clearly.  Maybe some day I'll learn this lesson.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>