Work on signal generation for simulation some more
It turns out that I was wrong about there being something wrong with the
ImSquared circuit: it seems to work just fine, and the fact that the
input signal gets used for both the blending and the multiplier input
appears perfectly ok.

The impedance into the multiplier should be very high as-is.  Famous
last confused words.

What confused me was misreading the output signal: I had expected hard
edges from the triangle input, and so I assumed that the soft curve was
due to some saturation or soft clipping issue and was blaming something
going wrong in the circuit.

But that was a complete thinko: the soft output curve was in fact just
the parabola of the triange wave (locally a linear "y=ax+b" function
when *not* looking at the peaks) being squared and becoming second-order
function ("y=az^2+b").

So those soft signals had nothing to do with the actual peaks of the
triangle - quite the opposite.  They were from the linear part of the
triangle wave when it switched signs.

Ergo, it actually worked exactly as designed, and I just misread the
output because I didn't line up the peaks and troughs correctly.

That said, the input signal is still much too simplistic.  The triangle
wave was better than a pure sine wave, but I want more dynamic range to
see any actual clipping issues.  So of the native signal generator
functions, let's try the amplitude-modulated signal instead, which gives
a combination of "quiet" and "loud" sections.

That already shows an interesting artifact: the output signal is
obviously a linear function of the current (it's the voltage drop over
the resistor R2), and I was naively expecting it to be centered around a
center voltage (the "half of the 1mA current" point, so roughly centered
at 8V in between the 7-9V peaks).

And it *is* technically centered right around the expected location, but
it's the *differential* output Vd(OUT1,OUT2) that is so centered.

And we obviously only tap one side (OUT1), and that's not centered
because that "x squared" function is obviously all one sign.  So you
actually get two disjoint OUT1/OUT2 voltage signals, and when you tap
only one side, you get this unbalanced thing.

The re-biasing to GNDREF then _slightly_ ends up also centering things,
but unless you are willing to lose a lot of low frequencies, not very
much.  So a more interesting amplitude-modulated input signal shows
these kinds of effects very clearly.

Will this matter? Probably not.  The signal might sound bad for any
number of reasons, and I doubt the partial biasing is a big issue.  But
this does make it clear that I need to use better input waveforms to see
issues like this.

The voltage-modulated signal is clearly a more interesting signal to
test.  There are probably better ones still, but let's at least take
that baby step.

You can change the amplitude numbers to actually see clear clipping
effects.  That's partially what the 'input' pot is supposed to be about,
but we have a definite voltage range for the input that the multiplier
biasing points work for, and that input range may be a bit too limited.

I clearly tuned this for the 0.1-0.2V amplitude range, and it shows.

Of course, maybe that "more distortion at higher signal amplitude
levels" is a good thing.  It's obviously how diode clipping works, so
it's certainly very traditional.  But I was kind of chasing the
distortion just from the frequency doubling part.

That said, I did hope that "small signals result in smaller distortion"
would make this multiplier a better-sounding octave-up effect than the
rectifier (along with the lack of introduced harsh edges at sign
crossing).

So "more distortion at high volume" _might_ be a feature, even if that
distortion then comes from the clipping in addition to the frequency
doubling.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>