RJ45: hopefully make the base board actually work
Ok, with some boards back and populated, that whole "feed signal over
RJ45" didn't work at all.  I still think the concept should work, but my
implementation of it is obviously very broken.

The op amp output to the effect ends up oscillating wildly at a
frequency of a few MHz.  And while I might have screwed something else
up, I think it's because of a capacitive load in the form of that cable.

According to random searches I did, the minimal fix is typically to add
a small resistive load in series on the output.  So I did that.

But thinking about it, the other end is very possibly going to have the
same problem, and none of the existing effects boards were designed with
any of this in mind.

So I will also be updating the test board at the other end of the cable
to do active buffering and I suspect I really should rebias the signal
too at that point so that the existing effects boards see the
traditional VGND setup instead of the quick hack that just passively
changed the rules under them.

It may obviously still be the case that this is all just misguided
garbage, but before throwing it in the bin, let's try to do it right.
That includes adding some transmission line simulation, even if that
didn't actually show the issue (with or without the series resistor).

Maybe it's something else entirely that I'm not seeing / understanding.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>