RJ45: update the "effect conversion" board to be active
... and hopefully make it work right.

The old hacky one was truly just a hack, and was a passive feed-through
of the signals from the RJ45 jack to the old effect board pin headers.

That was hacky mainly because it ignored the whole new world order wrt
voltages (ie the RJ45 cable carries dual voltage rails, the old effect
boards expect the VGND thing).

But it was also hacky because it expected the effects board to just deal
with the fact that the output now went to that RJ45 jack - and that
obviously didn't work very well for the base board, so why would it work
for the effects boards?

So this redoes the conversion board to actually buffer and rebias the
signals, and it basically looks a lot like the old VGND board except it
doesn't have a stomp switch or audio jacks, because all of that is
obviously on the new RJ45 base board.

I put all the components on the front, which in hindsight might have
been a mistake: it means that the joiner and splitter boards won't work
because they won't physically fit any more.  But if we want those in the
new RJ45 regime, I think they probably make more sense as actual new
boards anyway.

Will this work out ok? I still have no idea.  But nothing for it but
try.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>