ImSquared: fix enormously stupid mistake
Capacitors C1/C2 had been marked "Exclude from board", because they were
copied from a test capacitor I had in a sim-only circuit.  And I didn't
notice at any point that the PCB design didn't have them.

Not even when populating the board.  I did notice that the components on
the BOM started from C3, and I just idly thought - without looking any
closer at it - that I had forgotten to re-annotate the circuit.

So I never gave it another thought - until none of the signal actually
made it to the analog multiplier unit, so the board didn't actually work
at all.  D'oh!

Trivial fix, but oh how annoying.  KiCad makes it visually clear when a
component is excluded from simulation, and when a component is marked
"Do not populate".  But something marked "Exclude from board" is clearly
too easy to miss.  I need to be more careful.

I was able to bodge-wire this, with three of four endpoints of the two
mislaid caps being through-hole (the 'IN' signal source, and two
potentiometer pins).  For the fourth signal ('OUT1') I just soldered a
wire to the 2.2kOhm resistor.

I _hate_ soldering wire and my bodge wires weren't pretty.  But the
alternative was "wait a week for fixed boards to test because you're a
dum-dum".

The circuit does actually work and seems to behave exactly like in
simulation.  Lovely.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>