Add test boards for alternate connector setup
I'm still testing out alternatives to the "everything in one box" model.
This was really a result of me adding the buffering to the effect board
interface, and just being annoyed at the duplication of logic.

This one still has a CAT8 cable between the boxes, but makes the base
box purely passive, and adds USB-C as an alternative to RJ45.

USB-C actually has an analog audio standard ("Audio Accessory Mode").
This isn't quite it, because the voltage values are different and I
skipped the CCx resistors.  And I used left/right channel for in/out
rather than the microphone channel on SBUx lines.

So this is another horrid hack - but for "Linus testing stuff", maybe a
USB-C cable is even more convenient than ethernet?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>