Add 'discrete' base board
This is not me being old-fashioned as much as trying to actually just
use raw transistors rather than op-amps.  Not because it's better, but
because of the whole "maybe I'm actually starting to get transistors".

This still uses the TLE2426 for the virtual ground, so it's not really
discrete.  And I might end up looking for dual transistor packages.  Not
so much because of the matching (which was needed for the Gilbert cell,
but not so much for this kind of "use a single JFET as an amplifier) but
just to cut down on components a bit.

We'll see.

Anyway, this one is "tuned for simulation", with resistor values that
happen to give unity gain when using the spice values for the MMBFJ201
this uses.  That's just to verify that it's not distorting noticeably at
these voltages.  I'll add a pot or something if I actually make this a
board at some point.

In the meantime, this is just a "let's at least simulate a pure JFET
transistor front-end" thing, while I think about what form factors I
want to do etc (ie the whole "split up using a cable" thing).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>