Simplify the connector testing to just a USB-C connector
This is mainly to see what that would look like, particularly for the
base that now becomes really simple and small.

The effect connector ends up being crazy large just because of the old
pin header layout.  But the point of all this is that if it works out,
we'll dispense with the pin headers, and just use the USB-C connector
"natively".

So the effects board is just a temporary thing, and the base board is
where it's at.  If the base board works well, and I can stop soldering
stomp switches, I'll be good.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>