Update the simple zip-tie "lid" enclosure
I actually liked this for testing - it's not fancy, but it's just good
enough that you can pick up the board without worrying about affecting
the signal etc.  So I've been using this for rough testing when I don't
want to make it a pretty pedal, so let's just commit the slight updates
to the design.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>