ImSquared: do the output using a differential amplifier setup
Honestly, this doesn't really change anything fundamental, it's just
another step in the endless saga of "Linus does completely useless
things just because he wants to learn about circuits".  And building a
differential amplifier out of three op-amps is certainly that.

Doing this also seemed to show that the OPA1656 actually has an
appreciable amount of current going into the inputs in simulation.  It
may be a real effect when the input is close to the positive power rail
(which is outside spec for the chip).  Or it's a simulation issue.  Or
it's me messing something else up.

Repeat after me: "It's not a failure, it's a learning experience".

I'm not happy about the blending circuit, it annoys me how it shows up
in several of the effects circuits and it doesn't add anything to the
experience.  It's needed for testing actual sound, though, but it makes
me wonder if I should just do it in all in the base board (especially
the active RJ45 version?)

I also need to bring real sound into simulation.  I guess I will need to
learn about using wav files as inputs to ngspice, instead of hacking up
random signal sources just to see the effect in simulation.

Too many things to learn.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>