Try simulating an actual wav file input with new locally-built tooling
So this "works" in the sense that I built local copies of ngspice and
kicad and now it supports wav file input and output to simulation, and I
am trying to use some sample files to simulate what the heck it does and
sounds like.

That didn't work very well because the linear interpolation of samples
in the experimental ngspice wav branch seems quite broken, causing the
input signal to be more of a discrete staircase signal.

Which in turn then caused pretty much immediate "Timestep too small"
errors for the simulation because it's much too sensitive to not
converging, and the stepwise voltage jumps seemed to make an already
fragile situation much much worse.

I have a local hack for that ngspice sample interpolation issue that
makes the simulation mostly work, in that now I see the result on the
real signal.  I can make it even generate a result wav-file, but only if
I end up removing a lot of the surrounding circuitry, because ngspice
still ends up crapping out eventually even with my linearization fixlet.

The goal? See why the effect seems to simply not work very well on bass
guitar (although I have a suspicion that it's just the fact that the
rest of the sound path ends up filtering everything up and cleaning up
the fuzz).

I do suspect that the whole "use squaring as a octaver fuzz" is a
completely broken concept to begin with, but setting things up to use a
real signal is still a good idea.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>