Re-enable writing simulation result output wav file
Based on some hints from Holger Vogt I actually got the whole ImSquared
circuit to simulate for a full five-second sound simulation.  So now I
can hear it sounding bad without actually having hardware back.

So this enables the writeback of the sim-output.wav file, and uses the
offset parameter to get a proper signal output even at GNDREF instead of
having to re-bias the output to real ground in the circuit itself.

[ NOTE NOTE NOTE: I wasted more time than I'd like to admit on this due
  to interactions between kicad and ngspice.

  In particular, changing the parameters in the kicad text fields
  requires you to just quit and restart the whole thing for things to
  actually take effect.

  That was very confusing with me making changes to the parameters that
  didn't take effect. I'm sure I'll forget about this once more, so
  adding a big note to myself here about it. ]

It really doesn't sound great, but the circuit works, and simulation now
works too (but still needs my ngspice patch for input sample
interpolation).

The whole signal squaring approach does actually seem to work as a
octaver fuzz, but it's obviously a lot more sensitive to the input
voltage range than a plain precision rectification circuit is.

With the rectifier circuit, the output amplitude has a simple linear
relationship with the input signal, while the squaring function is
obviously non-linear.

It still might be an advantage once you mix in the original signal, with
a quiet signal giving much less distortion than a loud one.  But based
on limited testing, it ends up being pretty finicky so that "advantage"
is rather questionable.

Still, this exercise has taught me something new.  Both on the circuit
side and about ngspice.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>