Add discrete logic effect board for the passive USB-C base board
This board does two things:

 (a) it ends up being a test-board for the passive "just connect power
     and audio signals using USB-C" without using the conversion board

 (b) it's a random very old-fashioned "purely discrete logic" board

so this does away with op-amps and the virtual ground generation, by
virtue of being inexcusably stupid.  It's basically one single
transistor and a diode.  It is not a fancy board, in other words, but I
wanted to have some minimal test-board, and this was basically the
smallest thing I could make that still did _something_.

Using the old base board with virtual ground and op-amps would have made
this silly board even more pointless than it is right now, because this
circuit is so trivial that it doesn't actually want any op-amp
buffering, much less any virtual ground.

The sad part is that it probably sounds better than any of the fancier
boards.  It's the first one where I actually used real guitar sounds for
simulation (which is why the gitignore file also grows a "ignore mp3's
and wav files" to it).

But despite that it's all rather questionable, because who knows how
well simulation matches reality.  Particularly with a random JFET that
may end up having very different actual physical characteristics than
simulation had.

This is the kind of pedal that a sane person would do by soldering
components directly together.  But I'm doing the cray-cray "maybe I can
separate the stomp box from the effect with a USB cable" thing.  Because
the whole point is to play around with silly things.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>