Add symbol for NMB2227A NPN/PNP transistor pair
I've got matched pairs (NPN and PNP), but not a NPN/PNP combo symbol.
Since I've gone on a "discrete transistor" bender, but prefer doing the
"two transistors in a package" format, let's add this.

This was partly inspired by looking at a discrete rectifier circuit, but
I decided to go with a matched NPN pairing for that one.  So no direct
use for it yet, but maybe I'll do a discrete headphone amplifier with a
push-pull pair?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>