Add discrete rectifier circuit
All of my rectification circuits have sounded bad, and I should feel
bad.

Instead, I'm just making another one.  Because this one is different.

It's not different because it *sounds* any better.  No, it makes the
same horrible noise.

But it's different because it isn't based on some precision rectifier
circuit with an op-amp, but on a pair of transistors instead.  Loosely
based on Tim Escobedo's 'Push Me Pull You' pedal, except it's using a
matched NPN pair.

It can optionally amplify the out-of-phase signal to make it more of a
fuzzy thing.  Or something like that.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>