Convert the discrete test-boards to the new world order
The USB-C thing was good for testing things out, but not really
practical in real life.  So this just converts things to use the simpler
pin connector layout and the normal 1590A pot setup.

Both boards get a "brightness" pot, not because I think it makes much
sense, but because I standardized on the three-pot enclosure.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>