Random guitar pedal project gen3: Discrete Components
This is a reset of the 1590A pedal project, because unlike any sane
project, this one goes backwards.

Instead of making fancier and more complex pedals, this is the "Benjamin
Button" of guitar pedal projects and ages in reverse.  As the project
ages, it goes back in time and the circuits become more primitive.

So in Gen1, I was exploring power rails and packaging, with all the
signal paths being basically op-amp circuits.  In Gen2, I had largely
standardized on a virtual ground at 4.5V.

And then I was exploring long-tailed pairs and the like, and as I
started to feel like I understood some of the transistor circuits
better, I decided that I don't want to do op-amp, but regress to
entirely discrete components.

And that made the base board with signal buffering and virtual ground
actually do all these things that I didn't want my boards to do, so now
in Gen3 the base board becomes even more spartan, and the effect boards
have become minimalist.

Does any of this make any sense? No.  It never did as actual guitar
pedals that anybody would use.  But it was always about learning analog
circuits, and op-amps were a crutch.  Or training wheels, if you will.

Put another way: these gen3 circuits are not in any way better than gen2
(or even gen1).  But they are starting from a new base-line, and there's
no point in trying to claim there's a common history.

The previous circuits are still there in the git history, in the
'legacy' and 'legacy-gen2' branches.  As if anybody cared.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>