Fix up some initial Gen3 commit issues
Most notably, when I reset the history I effectively also screwed up the
credit for the FreeCAD enclosure, which isn't by yours truly.  So add an
explicit note on the history of that file with a pointer to the commit
in the history (see commit b5d59be: "Added FreeCAD native file of 1590A
enclosure").

Also fix up some paths for simulation workbooks that didn't get properly
updated when I renamed things by hand to remove the "Discrete" part of
the names.

And minor typo fix.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>