Clarify and tweak the clean boost circuit a bit
This adds more information about the actual transistor biasing choices,
and tweaks the details a bit more.  More importantly, this actually
documents the thing, with explanations for the particular choice of
resistors and voltage biases.

The maximum amplification is slightly less, but it should be pretty
clean, with a 15dB amplification of up to a 0.9V PtP signal (so turning
a 0.9V peak-to-peak signal into a 5V peak-to-peak one with no visible
clipping).

The frequency response should also be very flat.

Of course, that's all based on simulation.  As usual, reality may
disagree.  And note that the JFET input buffer adds its own limits, with
the J201 having a V_{GSoff} anywhere between -0.3 and -1.5.  So you
might end up clipping a 0.6V PtP input signal just due to that if you're
unlucky.

The simulation seems to be using a V_{GSoff} of around -0.8, so clipping
in simulation happens around 1.6V PtP even if the boost pot is set low.

And yes, I'm sure this could have been done with a single transistor,
but the JFET input buffer is for high input impedance, and the dual BJT
means that the output will be in the same phase as the input.  As if
that matters.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>