Add a three-way Bright / Clean / Blunt toggle to the boost circuit
Ok, I admit it: this was another thing mainly introduced because my
standard enclosure has three holes for pots and switches.

But it's also because "clean" is just too boring on an oscilloscope if
it actually ends up being as clean as it looks in simulation.

So the switch just adds options for bypassing the emitter resistor for
some not-so-clean boosts, where it's easier to get clipping to the point
of being almost a fuzz.

And hey, some people might want fuzz.

The basic clean boost was being pretty careful to have a flat frequency
response and should work fine at bass frequencies.  The emitter resistor
bypassing not so much.

Obviously the bright setting will boost higher frequencies (roughly
500+Hz), while the blunt one boosts most things over 100Hz - but not
really low bass.

Useful? If you have to ask, you clearly haven't been following this
project.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>