First revisions from gen3 testing...
It turns out that while the single connector was sufficient, it is also
mechanically annoyingly less robust than the two separate pin headers we
used to have.

So even though we don't need the extra pins, just split it up into
"power" and a "signal" connectors again.  The placement behind the
output jack was good, so just put the power jack behind the input one
instead.

That ends up making routing nicer with the power trace going down where
it needs to be for the LED anyway, so it's a win-win.  Much nicer than
the old board.  And with the extra pins and the position right next to
the LED, lets just make the base board LED voltage available as an
"active" signal in case anybody ever wants it.

Also, just move the DC blocking capacitor and the anti-pop resistor to
ground to the base board.  Any effects board needs to do that, so it's
one less thing to worry about.  Also add back the low pass RF filter.

With the connector change, this obviously makes the new boards
incompatible, but it's not like I have any legacy boards except for the
test-boards.

On the effect board side, this does the layout changes, but also fixes
the Octaver board that was half-arsed and hadn't done proper bias
voltage calculations.  This basically just reuses them from the boost
board.

Plus use a JFET input on for the Octaver, and while at it make it also
just split the phase the natural way.  Both simpler and more balanced.

So maybe the Octaver works now.

And by "works", I obviously don't mean "sounds good".

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>